// src/app/models/user.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { GrantedAuthority } from '../models/granted-authority';
import { Hackathon } from '../models/hackathon';
import { MonitorEvaluation } from '../models/monitor-evaluation';
import { Role } from '../models/role';
import { Skill } from '../models/skill';
import { SponsorApplication } from '../models/sponsor-application';
import { UserQuizScore } from '../models/user-quiz-score';
import { Workshop } from '../models/workshop';
export interface User {
  accountNonExpired?: boolean;
  accountNonLocked?: boolean;
  authorities?: Array<GrantedAuthority>;
  badge?: 'JUNIOR_COACH' | 'ASSISTANT_COACH' | 'SENIOR_COACH' | 'HEAD_COACH' | 'MASTER_MENTOR';
  birthdate?: string;
  createdAt?: string;
  credentialsNonExpired?: boolean;
  description?: string;
  email?: string;
  enabled?: boolean;
  evaluations?: Array<MonitorEvaluation>;
  hackathons?: Array<Hackathon>;
  id?: number;
  lastname?: string;
  monitorPoints?: number;
  name?: string;
  password?: string;
  picture?: string;
  roles?: Array<Role>;
  score?: number;
  skills?: Array<Skill>;
  sponsorApplication?: SponsorApplication;
  userQuizScores?: Array<UserQuizScore>;
  username?: string;
  workshops?: Array<Workshop>;
}